[fix] 영어 개행 문제 수정 및 맴버 아이디 반영
