+++ client/src/components/boards/BoardDetailAnswerCreate.js
@@ -1,0 +1,120 @@
+import styled from "styled-components";
+import axios from "axios";
+import { useForm } from "react-hook-form";
+import { useRecoilValue } from "recoil";
+import { boardState, memberIdState } from "../../states";
+
+// ! 나중에 서버 연결시 데이터 받아서 랜더링되도록 수정할 것
+const BoardDetailAnswerCreate = () => {
+    const board = useRecoilValue(boardState);
+    const memberId = useRecoilValue(memberIdState);
+    const url = "http://ec2-3-36-53-155.ap-northeast-2.compute.amazonaws.com:8080";
+
+    // 답글 등록 form
+    const {
+        register,
+        handleSubmit,
+        formState: { errors },
+    } = useForm();
+
+    // 답글 등록 요청 함수
+    const postComment = (data) => {
+        axios
+            .post(`/comments`, data)
+            .then((res) => {
+                window.location.reload();
+            })
+            .catch((err) => console.log(err));
+    };
+
+    return (
+        <BDAnswerCreateWrapper>
+            <BDAMainForm
+                onSubmit={handleSubmit((data) => {
+                    data.memberId = memberId;
+                    data.boardId = Number(board.boardId);
+                    postComment(data);
+                })}
+            >
+                <BDMainTextareaWrapper>
+                    <textarea
+                        placeholder="작성자에게 따듯한 응원과 격려를 보내주세요."
+                        className={errors.content ? "boardErrorTextarea1" : "boardTextarea1"}
+                        {...register("content", {
+                            required: true,
+                            minLength: 10,
+                        })}
+                    />
+
+                    <div>
+                        {errors.content ? <div className="boardErrorMessage">최소 10자 이상 작성해주세요.</div> : <div></div>}
+                        <button type="submit">답글 등록</button>
+                    </div>
+                </BDMainTextareaWrapper>
+            </BDAMainForm>
+        </BDAnswerCreateWrapper>
+    );
+};
+
+// styled components
+const BDAnswerCreateWrapper = styled.div`
+    width: 100%;
+    display: flex;
+    justify-content: center;
+`;
+
+// ------------- textarea wrapper ------------- //
+const BDAMainForm = styled.form`
+    padding: 30px;
+    width: 80%;
+    max-width: 1300px;
+
+    background-color: white;
+    box-shadow: 2px 2px 9px rgba(0, 0, 0, 0.5);
+    border-radius: 20px;
+
+    @media screen and (max-width: 768px) {
+        width: 100%;
+        padding: 20px;
+    }
+`;
+
+// ------------- textarea ------------- //
+const BDMainTextareaWrapper = styled.div`
+    textarea {
+        resize: none;
+        width: 100%;
+        max-width: 1300px;
+        height: 80px;
+        padding: 10px;
+
+        border: none;
+        border-radius: 5px;
+    }
+
+    div {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        margin-top: 3px;
+    }
+
+    button {
+        width: fit-content;
+        padding: 6px 10px;
+    }
+
+    @media screen and (max-width: 768px) {
+        div {
+            flex-direction: column;
+            align-items: flex-start;
+        }
+        button {
+            margin-top: 5px;
+            padding: 5px 6px;
+            font-size: 12px;
+        }
+    }
+`;
+
+export default BoardDetailAnswerCreate;
