+++ client/src/states/index.js
@@ -2,3 +2,4 @@ export * from "./boardState";
 export * from "./answerState";
 export * from "./selfCheckState";
 export * from "./selfCheckErrorState";
+export * from "./memberIdState";
